libc: Move transactional error handling into sub-directory error/
